Create ShipPlacement class
Created a new class that will store the
details related to a placement of a ship.

In general this works by starting at one
point on the board with a ship of known
length, determining valid possible placements
from there, and storing the indices to mark
as 'ship' if user selects the placement, as
well as the address of the 'endpoint' to
pass back so user can select one of the
possible placements that are legal based on
the present board state.