Modify Grid class ship tracking fields
Replaced a basic map of <cell-num, ship-name> with
the following :

map < cell-num, ship-num >
vec < Ship >
vec < ShipPlacement >

this way, per cell index, it is quick to identify
the name of the ship placed at that cell AND what
other cells the ship occupies via the placement.
the ship-num value above is simply the index of the
two listed vectors.